+++ Data/EventData.lua
@@ -239,7 +239,7 @@ end
 -- end
 
 function GetStartAndEndTime(secondsLeft, totalDuration) -- both in seconds
-    local endTime = floor((GetServerTime() + secondsLeft) / 3600 + 0.5) * 3600; -- Round to the hour
+    local endTime = floor((GetServerTime() + secondsLeft) / 300 + 0.5) * 300; -- Round to 5 minutes, 1 hour is not precise enough anymore
     local startTime = endTime - totalDuration;
 
     return startTime, endTime;