+++ GUI/WindowFrames/CategoriesFrame/CategoriesFrame.lua
@@ -11,9 +11,9 @@ function categoriesFrame:Load()
 	frame:SetPoint("RIGHT", AchievementFrameCategories, addon.Options.db.Window.CategoriesFrameWidthOffset, 0);
 
 	-- We need to insert the categories frame infront of the achievements frame so the show/hide function fire in the correct order
-	for i, frameName in next, ACHIEVEMENTFRAME_SUBFRAMES do
-		if frameName == addon.GUI.AchievementsFrame:GetName() then
-			tinsert(ACHIEVEMENTFRAME_SUBFRAMES, i, frame:GetName());
+	for i, subFrame in next, addon.GUI.SubFrames do
+		if subFrame == addon.GUI.AchievementsFrame then
+			tinsert(addon.GUI.SubFrames, i, frame);
 			break;
 		end
 	end
@@ -42,8 +42,8 @@ function KrowiAF_CategoriesFrame_OnShow(self) -- Used in Templates - KrowiAF_Cat
 	AchievementFrameCategories:Hide(); -- Issue #11: Fix
 	AchievementFrameFilterDropDown:Hide();
 	if not addon.IsWrathClassic then
-		AchievementFrame.searchBox:Hide();
-		AchievementFrameHeaderLeftDDLInset:Show();
+		AchievementFrame.SearchBox:Hide();
+		AchievementFrame.Header.LeftDDLInset:Show();
 	end
 	AchievementFrameWaterMark:SetTexture(addon.GUI.SelectedTab and addon.GUI.SelectedTab.WaterMark or "Interface/AchievementFrame/UI-Achievement-AchievementWatermark");
 	AchievementFrameCategoriesBG:SetTexCoord(0, 0.5, 0, 1); -- Set this global texture for player achievements
@@ -56,10 +56,10 @@ function KrowiAF_CategoriesFrame_OnHide() -- Used in Templates - KrowiAF_Categor
 	if not AchievementFrameAchievements:IsShown() then
 		AchievementFrameFilterDropDown:Hide();
 		if not addon.IsWrathClassic then
-			AchievementFrameHeaderLeftDDLInset:Hide();
+			AchievementFrame.Header.LeftDDLInset:Hide();
 		end
 	end
 	if not addon.IsWrathClassic then
-		AchievementFrame.searchBox:Show();
+		AchievementFrame.SearchBox:Show();
 	end
 end